fix(ui): implement useDebouncedCallback hook and integrate into SharePopover and useProjectLock
